Add some framework for loading data from MS files
Rows are loaded in chunks, so that memory usage is kept constant rather
than loading the whole file into memory. Chunking is implemented in the
storage backend so that it can take advantage of any tiling in the file
itself, and also amortise the costs of Python ops.

I tried using TaQL to select the channel and apply flagging, but it was
an order of magnitude slower than this approach.