Simplifications and generalizations of FFT code
Fftshift is simplified, eliminating the overly-complex alignment
argument, in favour of always accessing memory in chunks of the data
type.

The FFT is generalized to allow an arbitrary number of dimensions and to
support forward and inverse transforms. It now also relies on the
exact-padding branch in the cases where the padded size does not match
the size.