Support double-precision processing
The incoming visibilities, weights and UV coordinates are kept in single
precision, since that's plenty given the noise levels. Accumulation in
the grid is done in double precision, as is the FFT.

This is unlikely to get used in production; it's just handy to support to
determine whether accuracy bugs are really bugs or just the limits of
precision.

This also caught a number of places where numpy calculations were being
promoted from single to double precision because Python float constants
were present.