Use separability in convolutional gridding
The kernel was already modelled separably, but was previous multiplied
out for use in gridding. This change will significantly reduce memory
usage (particularly when many W planes are in place), at the expect of
more computation in the kernel. In the current implementation there are
also more memory loads, but that can be avoided at the expense of more
registers. Reducing the memory for the kernels will also assist with
read-only cache hit rates.

At the moment it is still not possible to do 64x64 kernels on K2100M due
to timeouts, which I'm assuming are due to register pressure/spilling.