Convert axis ordering to polarization, x, y
This puts all the data for one polarization together, which is important
to later memory-saving optimisations that process only one polarization
at a time and hence only need memory for one at a time. It may also
improve the memory access patterns for kernels that need to access all the
polarisations from a single thread.