Optimize gridding and degridding
Each block is now kept contiguous, instead of being split across the
window wrapping. This allows various coordinate calculations to be
hoisted outside the block loop.

To make this work, the kernel needs some zero padding, and hence the
default is reduced from 64 to 60.

A nice side effect is that the given --kernel-width is now exactly
respected (mathematically), rather than being increased to a power of 2.
This should make the GPU and CPU versions more similar.