First cut at implementing support for feed angle rotation
It is quite a lot slower because limitations in casacore mean that
the virtual columns need to be read one row at a time, instead of in
batches.

Because the calculations are per-visibility, all the polarization
matrix calculations are moved into the C++ code, using Eigen for the
linear algebra.