Work around a numpy 1.12 FutureWarning
It was complaining about using astype to convert from one structured
dtype to another, even though the behaviour under 1.13 would have been
the same (because the target has a prefix of the fields of the source).

To avoid it, a view is first created that interprets the source data
using the target dtype.