CLEAN to a threshold
The threshold is calculated as a fraction of the peak value at the start
of CLEAN, which I think is what wsclean is doing. The parameter for
number of minor cycles now becomes a maximum number, and is increased to
10000.

There are some issues at the edges of the image, due to the
antialiasing weight dropping to zero there and giving numeric
instability. This should be fixed in SR-289.

Closes SR-562.