Suppress katdal in requirements.txt if Python 3
Since ska-sa/katdal#150 katdal won't install on Python 3 anymore
(which makes sense as it is Python 2.7 only). The katsdpimager package
is tested on Python 3 as well, which then has broken requirements
because of this katdal restriction.

As katdal is used in `loader_katdal` in katsdpimager but not in the
tests, add a version qualifier to the requirement.