First stab at image-space continuum subtraction
It takes a local sky model (specified as a katpoint catalogue) and uses
it to initialise the model image. After completing the imaging, this
subtraction image is subtracted from the final model before applying the
restoring beam.

An issue that's emerged is that this works very well when the source is
on a pixel centre, but less well when it isn't, leading to side-lobes.
This can be improved with anti-aliasing. At the moment a simple 2x2
bilinear filter is used, but it can and should be improved in future
e.g. using the same AA filter used for gridding visibilities.