Gracefully handle completely flagged channels
Previously it would get a zero PSF and try to scale it up to have a peak
of 1.0, which lead to divide by zero and NaNs everywhere, which in turn
would crash the beam fitting.

Now if the PSF doesn't have a non-zero centre, it is assumed that all
the data has zero weight / is flagged.