Improve beam fitting
Previously a single threshold on the PSF to get data for both the
initial estimate and the fitted version. This lead to poor initial
estimates, which caused the non-linear least squares fit to go nuts.
It's now using only the bright stuff (> half power) to get the initial
estimate, which gives a much saner initial fit; the threshold for the
second phase can then be much lower and still give sensible results.

Closes SR-810.