Fix infinite loop in noise estimation
If there isn't enough data and the PSF is nasty then cleaning can get
into a positive feedback loop that can cause infinitely large values.
This in turn can cause the code that binary searches for the median to
get stuck in a loop. This lets it escape (the image will of course still
be garbage).