Move most of imager.py to new frontend.py
This is preparation for having common base code shared by the
command-line and pipeline versions of the imager. It also makes the code
easier to test, although that's not implemented yet.

The command-line option parsing is still in the script, because I
haven't decided exactly how to split that up.