Fix out-of-bounds memory accesses in predict kernel
The arrays are not padded to deal with the rounding in the number of
threads. That might be a better solution, but that would require some
refactoring. The changes to prevent out-of-bounds accesses are outside
the critical loops so it shouldn't affect performance much.

This fixes the crash in SPR1-298, but there is still another (unrelated)
test failure.