Fix calculation of density weights to account for padding
If the padding has non-zero values it would corrupt the calculation of
normalised thermal RMS by including it into the reductions.

This might not be an issue in practice (because the grid sizes will be
divisible by large powers of 2 to make them FFT-friendly, and hence are
unlikely to have any padding), but it was causing unit tests to fail on
my laptop. The test is extended to make the padding contents
reproducible.

This is the last fix needed for SPR1-298.