Discard visibilities with zero weight on any polarisation.
Zero weights generally indicates flagging, and if any
polarisation if flagged then at least one of the inputs is
contaminated and so no Stokes parameters will be clean.

The unit test needed changing because one of the test visibilities had
a single zero weight. I changed that weight, and added a new visibility
that I expect to be skipped.

I cheated on the new expected values in the test: I updated them to
match the actual values coming out after I changed the zero weight.

See SPR1-374.