Fix calculation of noise
When in SUMSQ clean mode, it was computing noise based on the median of
sqrt(I^2+Q^2+U^2+V^2) and scaling it as if it was the median of |I| to
get RMS noise. This means that it was reporting noise about 2.7x higher
than it should have (assuming I, Q, U, V noise is all independent, which
I'm not so sure about given that they're mixtures of H/V correlation
products).

The clean threshold calculation will also need to be fixed, as it will
also be cleaning deeper than it should.