Adjust calculation of noise threshold
When using CLEAN_SUMSQ, one needs to be slightly more conservative to
avoid cleaning the noise, because the metric for noise is larger than
for CLEAN_I. The approach taken is to adjust the scale factor so that
the probability of a noise pixel being cleaned remains the same.

This adds scipy to the requirements, but it's not really new because
it's indirectly required by katsdpsigproc.