Correct visibility-predicted RMS noise by sqrt(2)
katdal works with variance of amplitude, but equations for noise in an
image work with variance of the real/imaginary parts of the noise.

Also put in an abstraction so that we don't try to produce an estimate
for RMS noise for data loaded from measurement sets, where we have no
information about the scale of the weights.