Plot UV coverage in the report
It's added as an embedded SVG generated by matplotlib, rather than a
Bokeh plot, because Bokeh doesn't currently support elliptic arcs
(although matplotlib fakes them with Bezier curves, so it's possible
that this could be ported to Bokeh).

There may be some numerical stability issues for targets *exactly* on
the equatorial plane of date, because in that case the ellipse degrades
to a line. Targets with DEC=0 should generally be okay though because
that's the J2000 equator, which is not the same thing.