Checkpoint pipeline imaging state
If the pipelining imaging dies and is retried, detect channels that were
successfully imaged and skip them next time.

Also put in some defense against the case where the statistics are
written but the imager dies before reaching the checkpoint. This should
be less common now that the checkpoint" is a telstate key alongside the
other telstate keys (rather than successful process termination, which
can go wrong if Docker is having a bad hair day).

Closes SPR1-622.