Don't raise an exception if a Stopwatch is stopped twice
If an exception is thrown into a generator wrapped with
`@profile_generator` (in particular, if the generator is closed before
it finishes naturally), then the stopwatch will have been stopped inside
the context manager, and the context manager cleanup would try to stop
it again, raising an exception.

Instead, silently do nothing on a double stop.