Fix stack traces when using profile_generator
When the generator yielded, the timer was stopped, but it would still
appear in the stack trace if other profiled functions were called while
the generator was alive. I think this could severely mess up flamegraphs
because time spent in children is subtracted from total time to get time
spent only in self.

To fix this, pushing and popping the frame is now handled by the
stopwatch class on each stop and start.