Split out FixedImageParameters and FixedGridParameters
These are intended to capture information that is independent of the
channel being imaged. The eventual goal is to pass these rather than the
full ImageParameters/GridParameters in places that don't need to be
channel-specific (and from there eliminate multiple compilations of the
CUDA kernels), but this just applies the split.