Rewrite preprocessing compression
Instead of doing a sort by baseline (which proved to be somewhat
expensive), make it the caller's responsibility to provide
baseline-major data. For katdal that is relatively easy (just an axis
reordering). There is still a post-compression sort by w_slice, which is
necessary for batching data of the same w_slice (the HDF5 collector
performs poorly without this batching).

Some of the fields in vis_t were reordered so that the test for
mergability compares a contiguous part of the data structure.

The MS loader still need to be updated to do its own sorting.