Apply primary beam correction on the GPU
This should improve performance, and avoids moving the dirty image to
the CPU for correction then back to the GPU for convolution with the
restoring beam.

A side effect is that if the primary beam is written to file it will not
be clipped (as the clipping now happens on application).