Remove ability to pad FFTs along batch dimensions
This was never actually used (as far as I know), but was added at the
time because it seemed trivial to support. However, it was failing after
updating some code. My suspicion is that in certain cases, CUFFT is
doing some joint processing on several batches and then disentangling
them. When some of those batches contain uninitialised padding data,
this can lead to problem if the uninitialised data is very large
compared to the real data (or even Inf/NaN).