Rework handling of host buffers for transfers
Responsibility for copying data to the device has been moved from each
operation up to Imaging. This allows it to have just a single host
buffer per device buffer (with the possibility of double-buffering
later).

One remaining exception is sky models/images in the direct predictor,
which is still handled internally.

Imaging also takes care of synchronisation, which allows data to be
loaded from HDF5 and transferred to the HostArray buffers concurrently
with GPU work, which should help performance a little. queue.finish
calls in frontend.py are now just there to keep the progress bars
honest.