Downsample image before writing the thumbnail
This roughly halves the time taken to write the thumbnail. The resulting
image does not fairly different, because the noise is smoothed out and
this causes the colour scale to be tighter. That has the effect of
making the real flux stand out a bit better, so I'm happy with it.

Closes SPR1-793.