Rewrite primary_beam module in terms of katsdpmodels.primary_beam
This module pre-dated katsdpmodels.primary_beam and tried to present a
very general interface, but a different one to katsdpmodels. In future
katsdpmodels will be used for the primary beam models, and the
TrivialPrimaryBeam will be for backwards compatibility. It's thus
convenient for it to implement the common PrimaryBeam interface.

In theory it would now be possible to support any circularly-symmetric
PrimaryBeam, but since TrivialPrimaryBeam is the only such interface,
there is no point trying to accept arbitrary PrimaryBeam models until
there is support for models that rotate over time.

See SPR1-855.